Removed warnings

+++ Rekd-Engine/BezierCurve.cpp
@@ -17,7 +17,7 @@ Rekd2D::Core::Vector2F Rekd2D::Core::BezierCurve::GetPoint(float t)
 	for (int i = 0; i <= n; i++)
 	{
 		b = (float)((float)UtilMath::Factorial(n) / (float)(UtilMath::Factorial(i) * UtilMath::Factorial(n - i)));
-		res += b * powf(1 - t, n - i) * powf(t, i) * m_Points[i];
+		res += b * powf(1 - t, (float)(n - i)) * powf(t, (float)i) * m_Points[i];
 	}
 	return res;
 }
